createに使いたいから画像の部分をコンポーネント化
