+++ client/src/components/BokeImageCarousel.tsx/BokeImageCarousel.module.css
@@ -1,0 +1,22 @@
+.imageContainer {
+  position: absolute;
+  top: 350px;
+  display: flex;
+  width: 100%;
+  max-width: 1200px;
+  margin-top: 20px;
+  overflow: hidden;
+}
+
+.displayedImg {
+  flex: 1;
+  width: 10vw;
+  min-width: 100px;
+  max-width: 240px;
+  height: auto;
+  margin-right: 1vw;
+  margin-left: 1vw;
+  object-fit: cover;
+  background-color: #fff;
+  border: solid #434343;
+}
