+++ client/src/pages/index.module.css
@@ -46,25 +46,3 @@
   transform: translate(-75%, 50%);
 }
 
-.imageContainer {
-  position: absolute;
-  top: 350px;
-  display: flex;
-  width: 100%;
-  max-width: 1200px;
-  margin-top: 20px;
-  overflow: hidden;
-}
-
-.displayedImg {
-  flex: 1;
-  width: 10vw;
-  min-width: 100px;
-  max-width: 240px; /* 幅の上限 */
-  height: auto;
-  margin-right: 1vw;
-  margin-left: 1vw;
-  object-fit: cover;
-  background-color: #fff;
-  border: solid #434343;
-}
